+++ src/main/java/com/github/theredbrain/scriptblocks/registry/StatusEffectsRegistry.java
@@ -14,10 +14,10 @@ public class StatusEffectsRegistry {
 	public static final StatusEffect HOUSING_TRUSTED_EFFECT = new ScriptBlocksStatusEffect();
 	public static final StatusEffect HOUSING_GUEST_EFFECT = new ScriptBlocksStatusEffect();
 	public static final StatusEffect HOUSING_STRANGER_EFFECT = new ScriptBlocksStatusEffect();
-//	public static final StatusEffect EDIT_HOUSING_RESISTANCE_EFFECT = new ScriptBlocksStatusEffect();
 
 	public static final StatusEffect BUILDING_MODE = new ScriptBlocksStatusEffect();
 	public static final StatusEffect PORTAL_RESISTANCE_EFFECT = new ScriptBlocksStatusEffect();
+	public static final StatusEffect ADVENTURE_EFFECT = new ScriptBlocksStatusEffect();
 
 	public static void registerEffects() {
 		// --- Registration ---
@@ -26,9 +26,9 @@ public class StatusEffectsRegistry {
 		ScriptBlocks.HOUSING_TRUSTED_EFFECT = register("housing_trusted_effect", HOUSING_TRUSTED_EFFECT);
 		ScriptBlocks.HOUSING_GUEST_EFFECT = register("housing_guest_effect", HOUSING_GUEST_EFFECT);
 		ScriptBlocks.HOUSING_STRANGER_EFFECT = register("housing_stranger_effect", HOUSING_STRANGER_EFFECT);
-//		ScriptBlocks.EDIT_HOUSING_RESISTANCE_EFFECT = register("edit_housing_resistance_effect", EDIT_HOUSING_RESISTANCE_EFFECT);
 		ScriptBlocks.BUILDING_MODE = register("building_mode", BUILDING_MODE);
 		ScriptBlocks.PORTAL_RESISTANCE_EFFECT = register("portal_resistance_effect", PORTAL_RESISTANCE_EFFECT);
+		ScriptBlocks.ADVENTURE_EFFECT = register("adventure", ADVENTURE_EFFECT);
 	}
 
 	private static RegistryEntry<StatusEffect> register(String identifierString, StatusEffect statusEffect) {